feat(th3index): introduce temporal H3 cell index type
th3index is a first-class temporal type whose base domain is the H3
discrete global grid (Uber's hierarchical hexagonal tiling of the
Earth at resolutions 0–15).  Each instant stores a single H3 cell
identifier (a 64-bit integer).

The implementation covers the full MobilityDB type-integration checklist:

  Static base types
    h3index, h3indexset (h3indexspan and h3indexspanset exist for
    catalog completeness only — H3 cells have no meaningful total
    order so no span API is exposed to users)

  Temporal type
    th3index: all four subtypes — instant, discrete sequence, step
    sequence, linear sequence, and sequence set

  Spatial classification
    T_H3INDEX is a tspatial + tgeodetic type: SRID 4326, X=true,
    GEODETIC=true.  spatial_basetype(), spatial_flags(), spatial_srid(),
    and spatial_set_srid() all carry matching cases.

  tbigint ↔ th3index casts
    The casts use WITH FUNCTION wrappers (Tbigint_to_th3index /
    Th3index_to_tbigint) that rebuild TSequence/TSequenceSet from
    their instants so the STBox bbox is computed correctly for the
    target type.  TBox (tbigint layout) and STBox (th3index layout)
    have different sizes; TInstant has no embedded bbox and uses a
    type-tag rewrite.  set_basetype_quotes() treats T_H3INDEX values
    as unquoted integers.

  Invalid-cell bbox safety
    th3index_cell_set_stbox guards with isValidCell() before calling
    cellToBoundary.  For invalid cell IDs the H3 library returns
    E_SUCCESS with garbage coordinates; invalid cells produce a
    T-only STBox (X=false).

  Adjacent index recheck
    stbox_index_recheck requires a per-row recheck for
    RTAdjacentStrategyNumber.  adjacent_stbox_stbox returns TRUE for
    singleton periods contained within the query span, not only for
    geometrically adjacent pairs, so GiST/SP-GiST must not skip the
    recheck for this strategy.

  Lifted functions
    All h3-pg v1 functions lifted through the datum-wrapper /
    LiftedFunctionInfo machinery:
      - Inspection: getResolution, getBaseCellNumber, isValidCell,
        isResClassIII, isPentagon
      - Hierarchy: cellToParent, cellToCenterChild, cellToChildPos,
        childPosToCell (with next-resolution conveniences)
      - Lat/lng: latLngToCell, cellToLatLng (→tgeogpoint),
        cellToLatLng_tgeompoint (planar), cellToBoundary (→tgeography)
      - Directed edges: areNeighborCells, cellsToDirectedEdge,
        isValidDirectedEdge, origin/destination, boundary (→tgeography)
      - Vertices: cellToVertex, vertexToLatLng, isValidVertex
      - Grid traversal: gridDistance, cellToLocalIJ, localIJToCell
      - Metrics: cellArea, edgeLength, greatCircleDistance
        (km / m / rad / km² / m² / rad²)
    Datum wrappers live in h3index.c; hand-written GSERIALIZED adapters
    (h3_gs_point_to_cell, h3_cell_to_gs_point, h3_cell_to_gs_boundary,
    next-resolution helpers, IJ conversion, unit dispatch) live in the
    matching th3index_<topic>.c files and share declarations through
    meos/include/h3/th3index_internal.h.

  Set-valued static functions
    h3_grid_disk / h3_grid_ring / h3_grid_path_cells,
    h3_cell_to_children, h3_compact_cells, h3_uncompact_cells,
    h3_origin_to_directed_edges, h3_cell_to_vertexes,
    h3_get_icosahedron_faces — each returns h3indexset or intset.
    Temporal variants require a future tset<T> primitive and are
    deferred.

  Indexes
    GiST and SP-GiST on the time-only STBox bbox.

  Comparison operators and opclasses
    = ≠ < ≤ > ≥ lifted over the shared time axis; btree and hash
    opclasses for h3index.

  Documentation
    DocBook EN + ES chapters; Doxygen groups.

  Tests
    Test files 250–293 covering all subsystems; MEOS smoke harness;
    28/28 h3-family regression tests pass on PG 17.

210 MobilityDB regression tests pass.

+++ meos/src/temporal/spanset_meos.c
@@ -230,7 +230,7 @@ int_to_spanset(int i)
  * @csqlfn #Value_to_spanset()
  */
 SpanSet *
-bigint_to_spanset(int i)
+bigint_to_spanset(int64 i)
 {
   return value_spanset(i, T_INT8);
 }
